requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
djangorestframework-simplejwt
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.22.1,<0.23
requests>=2.28.2
django-guardian
django-rest-email-auth